fix(vaults): lint moved StakingVaults to services folder
